+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
-.
+
+## [8.6.2] - 2022-02-09
+
+-   Fixes #269. The json body parser will always be usd regardless of the fact whether the user has added the parser or not
 
 ## [8.6.1] - 2022-02-09